rename auth -> credential
