--- domain/repository/user/auth/auth.go
@@ -1,7 +1,0 @@
-package auth
-
-type IAuthRepository interface {
-	Append() error
-	Delete() error
-	IsExist() bool
-}
