+++ domain/repository/user/credential/credential.go
@@ -1,0 +1,7 @@
+package credential
+
+type ICredentialRepository interface {
+	Append() error
+	Delete() error
+	IsExist() bool
+}
